[Refactor]Routineコンポーネントの実装
